main: Move activateifneeded to own file
